Move file_pr to a central location, as it doesn't need to be duplicated.
